refactor(linker): isolate source discovery
