+++ src/linker/mod.rs
@@ -10,9 +10,14 @@ use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::rc::Rc;
-use walkdir::WalkDir;
 
 use crate::config::{Config, SyncType, TargetConfig};
+
+#[cfg(test)]
+use discovery::matches_path_glob;
+use discovery::matches_pattern;
+#[cfg(test)]
+use discovery::path_glob_match_iter;
 
 mod apply;
 mod clean;
@@ -445,301 +450,6 @@ impl Linker {
         Ok(())
     }
 
-    /// Expand a destination template for a single discovered file.
-    ///
-    /// Replaces the following placeholders:
-    /// * `{relative_path}` – parent directory of the matched file relative to
-    ///   the search root (e.g. `clients/agent-runtime`).  When the file is
-    ///   directly inside the search root, this is `.` (current directory).
-    /// * `{file_name}` – the file's name (e.g. `AGENTS.md`)
-    /// * `{stem}` – the file name without its extension (e.g. `AGENTS`)
-    /// * `{ext}` – the file's extension without the dot (e.g. `md`)
-    fn expand_destination_template(
-        template: &str,
-        rel_path: &Path, // path of the discovered file relative to search root
-    ) -> String {
-        // Optimization: use Cow<str> from to_string_lossy() to avoid heap allocations
-        // for valid UTF-8 paths.
-        let file_name = rel_path
-            .file_name()
-            .map(|n| n.to_string_lossy())
-            .unwrap_or_default();
-
-        let stem = rel_path
-            .file_stem()
-            .map(|s| s.to_string_lossy())
-            .unwrap_or_default();
-
-        let ext = rel_path
-            .extension()
-            .map(|e| e.to_string_lossy())
-            .unwrap_or_default();
-
-        // Use "." for files directly inside the search root so that
-        // templates like "{relative_path}/CLAUDE.md" produce a valid relative
-        // path ("./CLAUDE.md") rather than an absolute path ("/CLAUDE.md").
-        let dir = rel_path
-            .parent()
-            .map(|p| {
-                let s = p.to_string_lossy();
-                if s.is_empty() {
-                    std::borrow::Cow::Borrowed(".")
-                } else {
-                    s
-                }
-            })
-            .unwrap_or(std::borrow::Cow::Borrowed("."));
-
-        // Performance Optimization: single-pass expansion with String::with_capacity
-        // to eliminate O(N) heap allocations from multiple .replace() calls.
-        let mut result = String::with_capacity(template.len() + dir.len() + file_name.len());
-        let mut remaining = template;
-
-        while let Some(start_idx) = remaining.find('{') {
-            result.push_str(&remaining[..start_idx]);
-            let after_brace = &remaining[start_idx..];
-
-            if let Some(end_idx) = after_brace.find('}') {
-                let placeholder = &after_brace[..end_idx + 1];
-                match placeholder {
-                    "{relative_path}" => result.push_str(&dir),
-                    "{file_name}" => result.push_str(&file_name),
-                    "{stem}" => result.push_str(&stem),
-                    "{ext}" => result.push_str(&ext),
-                    _ => result.push_str(placeholder),
-                }
-                remaining = &after_brace[end_idx + 1..];
-            } else {
-                result.push('{');
-                remaining = &after_brace[1..];
-            }
-        }
-        result.push_str(remaining);
-        result
-    }
-
-    /// Process a `NestedGlob` target: walk `search_root`, match files against
-    /// `glob_pattern`, skip excluded paths, and create a symlink for each
-    /// Discovers files in a directory tree matching a glob pattern and creates
-    /// symlinks to each matched file.
-    ///
-    /// # Arguments
-    /// * `search_root` - Directory to walk for file discovery
-    /// * `glob_pattern` - Glob pattern with `**` support (e.g., `**/AGENTS.md`)
-    /// * `excludes` - Optional list of glob patterns to exclude from syncing
-    /// * `dest_template` - Destination path template supporting:
-    ///   - `{relative_path}` - Path relative to search root (`.` for root files)
-    ///   - `{file_name}` - Original filename with extension
-    ///   - `{stem}` - Filename without extension
-    ///   - `{ext}` - File extension (without leading dot)
-    /// * `options` - Sync options controlling verbose output and dry-run mode
-    ///
-    /// # Behavior
-    /// - Walks `search_root` recursively (following `follow_links = false` for safety)
-    /// - Skips directories and non-file entries
-    /// - Matches files against `glob_pattern` using `**` glob support
-    /// - Applies exclusion patterns if file matches any exclude, it is skipped
-    /// - Creates symlinks using `create_symlink()` which handles existing links
-    ///
-    /// # Performance Considerations
-    /// - Uses `find()` instead of `any()` for exclusion checks to enable early-exit
-    /// - Avoids exclusion iteration when `excludes` list is empty
-    ///
-    /// Internal helper to get cached nested-glob discovery results.
-    fn get_nested_glob_matches(
-        &self,
-        search_root: &Path,
-        glob_pattern: &str,
-        excludes: &[String],
-        options: &SyncOptions,
-    ) -> Result<NestedGlobMatches> {
-        let key = (
-            search_root.to_path_buf(),
-            glob_pattern.to_string(),
-            excludes.to_vec(),
-        );
-
-        let mut cache = self.glob_cache.borrow_mut();
-        if let Some(cached) = cache.get(&key) {
-            return Ok(Rc::clone(cached));
-        }
-
-        let mut found = Vec::new();
-        self.for_each_nested_glob_match(
-            search_root,
-            glob_pattern,
-            excludes,
-            options,
-            |full_path, rel_path| {
-                found.push((full_path.to_path_buf(), rel_path.to_path_buf()));
-                Ok(())
-            },
-        )?;
-        let rc_found = Rc::new(found);
-        cache.insert(key, Rc::clone(&rc_found));
-        Ok(rc_found)
-    }
-
-    fn process_nested_glob(
-        &self,
-        search_root: &Path,
-        glob_pattern: &str,
-        excludes: &[String],
-        dest_template: &str,
-        options: &SyncOptions,
-    ) -> Result<SyncResult> {
-        let mut result = SyncResult::default();
-
-        if !search_root.exists() || !search_root.is_dir() {
-            println!(
-                "  {} Search root does not exist: {}",
-                "!".yellow(),
-                search_root.display()
-            );
-            result.skipped += 1;
-            return Ok(result);
-        }
-
-        let matches = self.get_nested_glob_matches(search_root, glob_pattern, excludes, options)?;
-
-        for (full_path, rel_path) in matches.iter() {
-            let dest_str = Self::expand_destination_template(dest_template, rel_path);
-            if dest_str.is_empty() {
-                if options.verbose {
-                    println!(
-                        "  {} Destination template produced empty path for: {}",
-                        "!".yellow(),
-                        full_path.display()
-                    );
-                }
-                result.skipped += 1;
-                continue;
-            }
-
-            let dest = match self.ensure_safe_destination(&dest_str) {
-                Ok(dest) => dest,
-                Err(err) => {
-                    if options.verbose {
-                        println!(
-                            "  {} Skipping nested-glob destination {}: {}",
-                            "!".yellow(),
-                            dest_str,
-                            err
-                        );
-                    }
-                    result.skipped += 1;
-                    continue;
-                }
-            };
-
-            let resolved = ResolvedSource {
-                path: full_path.to_path_buf(),
-                exists: true,
-            };
-
-            let item_result = self.create_symlink(&resolved, &dest, options)?;
-            result.created += item_result.created;
-            result.updated += item_result.updated;
-            result.skipped += item_result.skipped;
-        }
-
-        Ok(result)
-    }
-
-    fn for_each_nested_glob_match<F>(
-        &self,
-        search_root: &Path,
-        glob_pattern: &str,
-        excludes: &[String],
-        options: &SyncOptions,
-        mut on_match: F,
-    ) -> Result<()>
-    where
-        F: FnMut(&Path, &Path) -> Result<()>,
-    {
-        // Optimization: Pre-split patterns once to avoid redundant allocations in the walk loop.
-        let split_pattern: Vec<&str> = glob_pattern.split('/').collect();
-        let split_excludes: Vec<Vec<&str>> =
-            excludes.iter().map(|e| e.split('/').collect()).collect();
-
-        let mut it = WalkDir::new(search_root).follow_links(false).into_iter();
-
-        while let Some(entry) = it.next() {
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(err) => {
-                    if options.verbose {
-                        let path = err
-                            .path()
-                            .map(|p| p.display().to_string())
-                            .unwrap_or_else(|| "<unknown path>".to_string());
-                        println!(
-                            "  {} WalkDir error while traversing {}: {}",
-                            "!".yellow(),
-                            path,
-                            err
-                        );
-                    }
-                    tracing::debug!(error = %err, path = ?err.path(), "WalkDir entry skipped during nested-glob traversal");
-                    continue;
-                }
-            };
-
-            let full_path = entry.path();
-            let rel_path = match full_path.strip_prefix(search_root) {
-                Ok(path) => path,
-                Err(_) => continue,
-            };
-
-            // Performance: generating a platform-agnostic relative path for glob matching.
-            // On Unix, this is zero-allocation (Cow::Borrowed).
-            // On Windows, it performs exactly one allocation for the backslash replacement.
-            let rel_os_str = rel_path.to_string_lossy();
-            let rel_str = if std::path::MAIN_SEPARATOR == '/' {
-                rel_os_str
-            } else {
-                std::borrow::Cow::Owned(rel_os_str.replace(std::path::MAIN_SEPARATOR, "/"))
-            };
-
-            if rel_str.is_empty() {
-                continue;
-            }
-
-            let path_it = rel_str.split('/');
-            if let Some(idx) = split_excludes
-                .iter()
-                .position(|exclude_parts| path_glob_match_iter(path_it.clone(), exclude_parts))
-            {
-                let matched_exclude = &excludes[idx];
-                if options.verbose {
-                    println!(
-                        "  {} Excluded by '{}': {}",
-                        "○".yellow(),
-                        matched_exclude,
-                        full_path.display()
-                    );
-                }
-
-                if entry.file_type().is_dir() {
-                    it.skip_current_dir();
-                }
-                continue;
-            }
-
-            if !entry.file_type().is_file() {
-                continue;
-            }
-
-            if !path_glob_match_iter(path_it, &split_pattern) {
-                continue;
-            }
-
-            on_match(full_path, rel_path)?;
-        }
-
-        Ok(())
-    }
-
     /// Process a `module-map` target: iterate mappings and create a symlink
     /// for each one, resolving the destination filename from:
     /// 1. mapping.filename_override (explicit user choice)
@@ -1164,143 +874,6 @@ fn normalize_inline_whitespace_to(line: &str, out: &mut String) {
             out.push(ch);
         }
     }
-}
-
-/// Simple glob pattern matching (supports * and ?)
-/// This is an iterative implementation which is more performant than the previous
-/// recursive one, especially for patterns with '*' since it avoids string
-/// allocations and recursion. It uses backtracking with a stored star position.
-/// This implementation is Unicode-aware.
-fn matches_pattern(name: &str, pattern: &str) -> bool {
-    let mut name_it = name.chars();
-    let mut pattern_it = pattern.chars();
-
-    let mut star_p_it = None;
-    let mut star_n_it = None;
-
-    loop {
-        let s_char = name_it.clone().next();
-        let p_char = pattern_it.clone().next();
-
-        match (s_char, p_char) {
-            (Some(s), Some(p)) if p == s || p == '?' => {
-                name_it.next();
-                pattern_it.next();
-            }
-            (_, Some('*')) => {
-                pattern_it.next();
-                star_p_it = Some(pattern_it.clone());
-                star_n_it = Some(name_it.clone());
-            }
-            (Some(_), _) => {
-                // Name has chars, but pattern doesn't match and is not '*'
-                if let (Some(star_p), Some(star_n)) = (star_p_it.as_mut(), star_n_it.as_mut()) {
-                    if star_n.next().is_none() {
-                        return false;
-                    }
-                    name_it = star_n.clone();
-                    pattern_it = star_p.clone();
-                } else {
-                    return false; // Mismatch and no star to backtrack to.
-                }
-            }
-            (None, _) => {
-                // Name is exhausted
-                return pattern_it.all(|c| c == '*');
-            }
-        }
-    }
-}
-
-/// Path-aware glob pattern matching that supports the `**` double-star wildcard.
-///
-/// Rules:
-/// * `**` (as a whole path segment) matches any number of path segments,
-///   including zero segments.
-/// * `*` within a segment matches any sequence of characters that does not
-///   include a `/`.
-/// * `?` within a segment matches any single character other than `/`.
-///
-/// The `path` argument must use `/` as the path separator.  Use
-/// [`matches_pattern`] for single-segment (filename-only) matching.
-#[cfg(test)]
-fn matches_path_glob(path: &str, pattern: &str) -> bool {
-    let pattern_parts: Vec<&str> = pattern.split('/').collect();
-    path_glob_match_iter(path.split('/'), &pattern_parts)
-}
-
-/// Core path-aware glob matching logic using iterators to avoid allocations.
-/// This implementation uses a backtracking algorithm which is more performant than a
-/// recursive one, especially for patterns with '**' since it avoids stack overhead.
-/// It provides O(N*M) complexity in typical cases.
-fn path_glob_match_iter<'a, I>(mut path_it: I, pattern: &[&str]) -> bool
-where
-    I: Iterator<Item = &'a str> + Clone,
-{
-    let mut pat_idx = 0;
-    let mut backtrack_path_it: Option<I> = None;
-    let mut backtrack_pat_idx = None;
-
-    loop {
-        let mut path_it_peek = path_it.clone();
-        match path_it_peek.next() {
-            Some(s) => {
-                if !try_match_segment(
-                    s,
-                    &mut path_it,
-                    pattern,
-                    &mut pat_idx,
-                    &mut backtrack_path_it,
-                    &mut backtrack_pat_idx,
-                ) {
-                    return false;
-                }
-            }
-            None => {
-                // Path is exhausted. Remaining pattern must be all "**"
-                while pat_idx < pattern.len() && pattern[pat_idx] == "**" {
-                    pat_idx += 1;
-                }
-                return pat_idx == pattern.len();
-            }
-        }
-    }
-}
-
-/// Process a single path segment against the current pattern state.
-/// Returns false if matching definitively fails, true to continue.
-fn try_match_segment<'a, I>(
-    segment: &str,
-    path_it: &mut I,
-    pattern: &[&str],
-    pat_idx: &mut usize,
-    backtrack_path_it: &mut Option<I>,
-    backtrack_pat_idx: &mut Option<usize>,
-) -> bool
-where
-    I: Iterator<Item = &'a str> + Clone,
-{
-    if *pat_idx < pattern.len() && pattern[*pat_idx] == "**" {
-        // ** matches zero or more segments
-        *backtrack_pat_idx = Some(*pat_idx);
-        *backtrack_path_it = Some(path_it.clone());
-        *pat_idx += 1;
-    } else if *pat_idx < pattern.len() && matches_pattern(segment, pattern[*pat_idx]) {
-        path_it.next(); // Consume segment
-        *pat_idx += 1;
-    } else if let (Some(b_pat_idx), Some(b_path_it)) =
-        (*backtrack_pat_idx, backtrack_path_it.as_mut())
-    {
-        // Backtrack: last ** matches one more segment
-        if b_path_it.next().is_none() {
-            return false;
-        }
-        *path_it = b_path_it.clone();
-        *pat_idx = b_pat_idx + 1;
-    } else {
-        return false;
-    }
-    true
 }
 
 #[cfg(test)]
